+++ lib/pow/store/credentials_cache.ex
@@ -12,6 +12,10 @@ defmodule Pow.Store.CredentialsCache do
     * `users/2` - to list all current users
     * `sessions/2` - to list all current sessions
 
+  The `:ttl` should be maximum 30 minutes per
+  [OWASP recommendations](https://cheatsheetseries.owasp.org/cheatsheets/Session_Management_Cheat_Sheet.html#session-expiration).
+  A warning will be output for any sessions created with a longer TTL.
+
   ## Custom credentials cache module
 
   Pow may use the utility methods in this module. To ensure all required
@@ -42,11 +46,14 @@ defmodule Pow.Store.CredentialsCache do
       from the context. Defaults false.
 
   """
-  alias Pow.{Operations, Store.Base}
+  alias Pow.{Config, Operations, Store.Base}
 
   @callback users(Base.config(), module()) :: [any()]
   @callback sessions(Base.config(), map()) :: [binary()]
   @callback put(Base.config(), binary(), {map(), list()}) :: :ok
+
+  # Per https://cheatsheetseries.owasp.org/cheatsheets/Session_Management_Cheat_Sheet.html#session-expiration
+  @recommended_max_idle_timeout :timer.minutes(30)
 
   use Base,
     ttl: :timer.minutes(30),
@@ -111,7 +118,23 @@ defmodule Pow.Store.CredentialsCache do
 
     delete_user_sessions_with_fingerprint(config, user, metadata)
 
-    Base.put(config, backend_config(config), records)
+    backend_config =
+      config
+      |> backend_config()
+      |> warn_maximum_timeout()
+
+    Base.put(config, backend_config, records)
+  end
+
+  defp warn_maximum_timeout(config) do
+    if Config.get(config, :ttl, 0) > @recommended_max_idle_timeout do
+      IO.warn(
+        """
+        warning: `:ttl` value for sessions should be no longer than #{round(@recommended_max_idle_timeout / 1_000 / 60)} minutes to prevent session hijack, please consider lowering the value
+        """)
+    end
+
+    config
   end
 
   @doc """