+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## v1.0.24 (TBA)
+## v1.0.24 (2021-05-27)
 
 ### Enhancements